+++ core/src/main/kotlin/io/customer/android/sdk/CustomerIOInstance.kt
@@ -1,0 +1,7 @@
+package io.customer.android.sdk
+
+/**
+ * Allows mocking of [CustomerIO] for your automated tests in your project. Mock [CustomerIO] to assert your code is calling functions
+ * of the SDK and/or do not have the SDK run it's real implementation during automated tests.
+ */
+interface CustomerIOInstance
