--- core/src/main/kotlin/io/customer/sdk/CustomerIOInstance.kt
@@ -1,3 +1,0 @@
-package io.customer.sdk
-
-interface CustomerIOInstance
